Fix #songs test to add song instances to @song
The #songs spec is testing for the @songs array to show Song instances within an Artist's collection. However, the test does not add instances yet expects instances to be present. The failure is confusing students who have otherwise functional code.

This commit fixes the code to add instances via the #add_song method and also reorders the #songs test to come after the #add_song test for easier comprehension.